requirements.txt
alembic==1.14.1
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.1.31
click==8.1.8
distlib==0.3.9
filelock==3.16.1
flasgger==0.9.7.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-serialize==2.2.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==21.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.9
MarkupSafe==2.1.5
mistune==3.1.3
packaging==24.2
permissive-dict==1.0.4
pipenv==2024.4.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
SQLAlchemy==2.0.39
SQLAlchemy-serializer==1.4.12
typing_extensions==4.12.2
virtualenv==20.29.3
Werkzeug==3.0.6
zipp==3.20.2
